campus-genius-backend/requirements.txt
# Django and DRF
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.1
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
dj-rest-auth[with_social]
django-allauth==0.58.2

# Database
mysqlclient==2.2.4
dj-database-url==2.1.0

# FastAPI and AI Service
fastapi==0.109.1
uvicorn==0.27.0
python-socketio==5.11.1
python-engineio==4.8.1
pymongo==4.6.1
python-dotenv==1.0.0
motor==3.3.2
python-multipart==0.0.6
pydantic==2.6.1
openai==1.12.0
langchain==0.1.4
langchain-google-genai==0.0.9
pinecone-client==3.0.2

# Testing and Development
pytest==8.0.0
pytest-django==4.7.0
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Utilities
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2 
